+++ skrub/__init__.py
@@ -5,7 +5,6 @@ skrub: Prepping tables for machine learning.
 from pathlib import Path as _Path
 
 from . import _selectors as selectors
-from ._adaptive_squashing import AdaptiveSquashingTransformer
 from ._agg_joiner import AggJoiner, AggTarget
 from ._check_dependencies import check_dependencies
 from ._column_associations import column_associations
@@ -38,6 +37,7 @@ from ._multi_agg_joiner import MultiAggJoiner
 from ._reporting import TableReport, patch_display, unpatch_display
 from ._select_cols import Drop, DropCols, SelectCols
 from ._similarity_encoder import SimilarityEncoder
+from ._squashing_scaler import SquashingScaler
 from ._string_encoder import StringEncoder
 from ._table_vectorizer import Cleaner, TableVectorizer
 from ._tabular_learner import tabular_learner
@@ -99,5 +99,5 @@ __all__ = [
     "StringEncoder",
     "column_associations",
     "toy_orders",
-    "AdaptiveSquashingTransformer",
+    "SquashingScaler",
 ]